% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/product_data.R
\name{add_product_colours}
\alias{add_product_colours}
\title{Add Atlas of Economic Complexity section colours to products}
\usage{
add_product_colours(data, digits)
}
\arguments{
\item{data}{a data frame with a column of hs92 product codes}

\item{digits}{representation of hs92 product codes. either four or six digits.}
}
\value{
tibble
}
\description{
Add Atlas of Economic Complexity section colours to products
}
\examples{
\dontrun{
library(complexitydata)
state_economic_complexity |> 
add_product_colours(digits = 4)
}
}
