% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_product_names.R
\name{add_product_names}
\alias{add_product_names}
\title{Add product names to a dataset}
\usage{
add_product_names(data, digits)
}
\arguments{
\item{data}{}

\item{digits}{}
}
\description{
Add product names to a dataset
}
