% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_graphs.R
\name{atlas_complexity_colours}
\alias{atlas_complexity_colours}
\title{Atlas of Economic Complexity PCI colour map}
\usage{
atlas_complexity_colours()
}
\description{
Atlas of Economic Complexity PCI colour map
}
