% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlas_countries}
\alias{atlas_countries}
\title{Atlas of Economic Complexity Countries.}
\format{
A tibble with 133 rows and 2 columns:
\describe{
\item{Country}{Country}
\item{location_code}{3-digit ISO country code}
}
}
\usage{
atlas_countries
}
\description{
The name and ISO-3 representation of countries included in the Atlas of Economic Complexity
}
\keyword{datasets}
