% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexity.R
\name{calculate_complexity}
\alias{calculate_complexity}
\title{Calculate Complexity Measures}
\usage{
calculate_complexity(data, year)
}
\arguments{
\item{data}{export data}

\item{year}{year to calculate complexity metrics}
}
\value{
dataframe
}
\description{
Calculate Complexity Measures
}
