% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexity.R
\name{calculate_complexity_shares}
\alias{calculate_complexity_shares}
\title{Calculate economic complexity indicators using Davies and Mare 2021 method of weighted correlations.}
\usage{
calculate_complexity_shares(data, region, product, value, verbose = FALSE)
}
\arguments{
\item{data}{data suitable for calculating complexity.}

\item{region}{variable in \code{data} which corresponds to a region.}

\item{product}{variable in \code{data} which corresponds to a product.}

\item{value}{variable in \code{data} which corresponds to a value.}

\item{verbose}{logical. FALSE (the default) provides no messages to the user.}

\item{method}{the method used to calculate economic complexity indicators. One of 'eigenvalue', 'fitness', or 'shares'.}
}
\value{
the supplied \code{data} with calculated city and activity complexity.
}
\description{
Calculate economic complexity indicators using Davies and Mare 2021 method of weighted correlations.
}
\examples{
\dontrun{

calculate_complexity_shares(data, region = "sa3", product = "indp", value = "count")

}

}
