% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexity.R
\name{calculate_complexity_time_series}
\alias{calculate_complexity_time_series}
\title{Calculate complexity (multiple years)}
\usage{
calculate_complexity_time_series(data, years)
}
\arguments{
\item{data}{export data}

\item{year}{years to calculate complexity metrics}
}
\value{
dataframe
}
\description{
Calculate complexity (multiple years)
}
