% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexity.R
\name{classify_products}
\alias{classify_products}
\title{Classify complexity products}
\usage{
classify_products(complexity_data, base, current, country = NULL)
}
\arguments{
\item{complexity_data}{data with complexity measures}

\item{base}{base year for comparison}

\item{current}{current year for comparison}

\item{country}{country for comparison}
}
\value{
dataframe
}
\description{
Classify complexity products
}
