% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complexity_classification}
\alias{complexity_classification}
\title{Atlas of Economic Complexity colours and sectors.}
\format{
A tibble with 1,217 rows and 3 columns:
\describe{
\item{hs_product_code}{character representation of the product code - HS92 4 digits}
\item{sector}{character representation of the Atlas sector}
\item{colour}{hex code of the colour used on the Atlas website for each sector}
}
}
\source{
\url{https://doi.org/10.7910/DVN/3BAL1O}
}
\usage{
complexity_classification
}
\description{
The Atlas of Economic Complexity mapping between 4-digit HS92 product codes and industry sectors.
The specified colour is equivalent to the colours used on the Atlas website.
}
\keyword{datasets}
