% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_product_space.R
\name{create_product_space}
\alias{create_product_space}
\title{Title}
\usage{
create_product_space(data, country, year, services = FALSE)
}
\arguments{
\item{country}{}

\item{year}{}
}
\description{
Title
}
