% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecomplexity-package.R
\docType{package}
\name{ecomplexity-package}
\alias{ecomplexity}
\alias{ecomplexity-package}
\title{ecomplexity: Functions for Common Economic Complexity Operations}
\description{
This package contains helper functions for some commonly requested Economic Complexity analysis.
}
\author{
\strong{Maintainer}: Hamish Gamble \email{hamish.gamble@flinders.edu.au} (\href{https://orcid.org/0000-0003-1277-1227}{ORCID})

}
\keyword{internal}
