% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_graphs.R
\name{graph_complexity_coverage}
\alias{graph_complexity_coverage}
\title{Data Coverage}
\usage{
graph_complexity_coverage(data, region, activity, flip = FALSE)
}
\arguments{
\item{data}{economic complexity input data}

\item{region}{quoted name of the region}

\item{activity}{quoted name of the 'activity' or product}

\item{flip}{logical to flip the graph}
}
\value{
ggplot2 object
}
\description{
Data Coverage
}
\examples{
\dontrun{
graph_complexity_coverage(data, "sa3", "indp")
}
}
