% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_graphs.R
\name{graph_complexity_map}
\alias{graph_complexity_map}
\title{Complexity map}
\usage{
graph_complexity_map(data, fill.var)
}
\arguments{
\item{data}{data of class sf}

\item{fill.var}{quoted name of the variable used for the fill aesthetic}
}
\value{
tmap object
}
\description{
Complexity map
}
\examples{
\dontrun{
graph_complexity_map(data, "city_complexity_index")}
}
