% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_graphs.R
\name{graph_complexity_opportunities}
\alias{graph_complexity_opportunities}
\title{Complexity opportunities}
\usage{
graph_complexity_opportunities(data, region, industry)
}
\arguments{
\item{data}{}

\item{region}{}

\item{industry}{}
}
\value{
ggplot
}
\description{
Complexity opportunities
}
