% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_graphs.R
\name{graph_complexity_product_space}
\alias{graph_complexity_product_space}
\title{Draw Product Spaces}
\usage{
graph_complexity_product_space(country, year, services = FALSE)
}
\arguments{
\item{country}{country}

\item{year}{year}

\item{services}{should services be included? Default is FALSE which excludes service exports}
}
\value{
ggplot
}
\description{
Draw Product Spaces
}
\examples{
graph_complexity_product_space("AUS",)
}
