% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_graphs.R
\name{graph_complexity_rank}
\alias{graph_complexity_rank}
\title{Economic complexity rank time series}
\usage{
graph_complexity_rank(data)
}
\arguments{
\item{data}{complexity dataset from \code{complexitydata} package}
}
\value{
ggraph2 object
}
\description{
Economic complexity rank time series
}
\examples{
\dontrun{
library(complexitydata) 
graph_complexity_rank(data = state_economic_complexity)
}
}
