% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_graphs.R
\name{graph_complexity_tree}
\alias{graph_complexity_tree}
\title{Export Tree Map.}
\usage{
graph_complexity_tree(data, year, region, classification)
}
\arguments{
\item{data}{export data}

\item{year}{year}

\item{region}{which region to draw the map.}

\item{classification}{hs92/hs12/hs22}
}
\value{
ggplot2 object
}
\description{
Colours adjusted to match Atlas of Economic Complexity.
}
\examples{
data <- read_complexitydata("state_economic_complexity")
graph_complexity_tree(data, 2022, "SA", classification = "hs92")
}
