% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_colours}
\alias{product_colours}
\title{Harmonised System (1992) product colours.}
\format{
A tibble with 1,241 rows and 3 columns:
\describe{
\item{hs4}{character representation of the product code - HS92 4 digits}
\item{color}{hex code of product colour}
\item{product_id}{id}
}
}
\usage{
product_colours
}
\description{
Colour groupings as used by the Atlas of Economic Complexity publication for product spaces.
}
\keyword{datasets}
