% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_data}
\alias{product_data}
\title{Harmonised System (1992/2012/2022) product data.}
\format{
Tibble with
\describe{
\item{code_6}{Six digit product code}
\item{code_4}{Four digit product code}
\item{code_2}{Two digit product code}
\item{code_1}{One digit product code}
\item{name_6}{Six digit product name}
\item{name_4}{Four digit product name}
\item{name_2}{Two digit product name}
\item{name_1}{One digit product name}
\item{classification}{HS Classification}
}
}
\usage{
product_data
}
\description{
A dataset containing the names, product code, group name and section name of products
the Harmonised System.
}
\keyword{datasets}
