% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_data12}
\alias{product_data12}
\title{Harmonised System (2012) product data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5207 rows and 9 columns.
}
\usage{
product_data12
}
\description{
Harmonised System (2012) product data
}
\keyword{datasets}
