% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_data_all}
\alias{product_data_all}
\title{Harmonised System (1992) product data.}
\format{
A tibble with 1,248 rows and 6 columns:
\describe{
\item{product_id}{id}
\item{parent_id}{link to products parent group}
\item{hs_product_name_short_en}{english short name of the product}
\item{hs_product_code}{character representation of the product code - HS92 4 digits}
\item{group_name}{english short name of the product group}
\item{section_name}{english short name of the product section}
}
}
\source{
\url{https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/3BAL1O/WQZPYX&version=4.0}
}
\usage{
product_data_all
}
\description{
A dataset containing the names, product code, group name and section name of products
categorised under the 1992 version of the Harmonised System.
}
\keyword{datasets}
