% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_api.R
\name{read_complexity_api}
\alias{read_complexity_api}
\title{Read data from the Economic Complexity API}
\usage{
read_complexity_api(query, query_name)
}
\arguments{
\item{query}{GraphiQL query. See \verb{https://github.com/harvard-growth-lab/api-docs/blob/main/atlas.md} for examples.}

\item{query_name}{The name of the query. Whatever you want to call it.}
}
\value{
dataframe
}
\description{
Read data from the Economic Complexity API
}
\examples{
\notrun{}
}
