% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_complexitydata.R
\name{read_complexitydata}
\alias{read_complexitydata}
\title{Read and load into the environment, economic complexity data from the complexitydata GitHub repository}
\usage{
read_complexitydata(name = NULL, refresh = FALSE, export_dir = tempdir())
}
\arguments{
\item{name}{the name of the dataset available at \code{aiti-flinders/complexitydata}}

\item{refresh}{TRUE to force re download the file, even if it already exists on disk.}

\item{export_dir}{directory to save the downloaded file. The default saves to a temporary directory.}
}
\value{
the downloaded data is loaded into the current R environment.
}
\description{
Read and load into the environment, economic complexity data from the complexitydata GitHub repository
}
\examples{
\dontrun{
read_complexitydata("sa3_indp1_occp1")
}
}
