% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_complexitydata.R
\name{read_complexitydata_p}
\alias{read_complexitydata_p}
\title{Read complexity parquet files}
\usage{
read_complexitydata_p(
  classification,
  digits,
  refresh = FALSE,
  export_dir = tempdir()
)
}
\arguments{
\item{classification}{data classification}

\item{digits}{data digits}

\item{refresh}{force redownload}

\item{export_dir}{export directory}
}
\value{
parquet
}
\description{
Read complexity parquet files
}
\examples{
\dontrun{
read_complexity_p("hs92", "4")}
}
