% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_product_data.R
\name{add_product_names}
\alias{add_product_names}
\title{Add product names to a dataset}
\usage{
add_product_names(data, digits)
}
\arguments{
\item{data}{a data frame with a column of hs92 product codes}

\item{digits}{representation of hs92 product codes. either four or six digits.}
}
\value{
tibble
}
\description{
Add product names to a dataset
}
\examples{
\dontrun{
library(complexitydata)
state_economic_complexity |> 
add_product_digits(digits = 4)
}
}
