% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anzsic_hs}
\alias{anzsic_hs}
\title{ANZSIC to Harmonised System concordance}
\format{
A tibble with 1,521 rows and 2 columns:
\describe{
\item{hs_product_code}{character representation of the product code - HS92 4 digits}
\item{anzsic}{character representation of the ANZSIC class}
}
}
\usage{
anzsic_hs
}
\description{
The relationship between product codes and Australian industry classifications.
}
\keyword{datasets}
