% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexity.R
\name{calculate_complexity}
\alias{calculate_complexity}
\title{Calculate economic complexity indicators}
\usage{
calculate_complexity(data, region, product, value, year)
}
\arguments{
\item{data}{data suitable for calculating complexity.}

\item{region}{variable in \code{data} which corresponds to a region.}

\item{product}{variable in \code{data} which corresponds to a product.}

\item{value}{variable in \code{data} which corresponds to a value.}

\item{year}{year to calculate complexity indicators.}
}
\value{
dataframe
}
\description{
Calculate economic complexity indicators
}
\examples{
\dontrun{

calculate_complexity(state_data,
region = "location_code",
product = "hs_product_code",
value = "export_value",
year = 1996)

}
}
