% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complexity.R
\name{calculate_complexity_time_series}
\alias{calculate_complexity_time_series}
\title{Calculate economic complexity indicators for multiple years}
\usage{
calculate_complexity_time_series(data, years, region, product, value)
}
\arguments{
\item{data}{data suitable for calculating complexity.}

\item{years}{years to calculate complexity indicators.}

\item{region}{variable in \code{data} which corresponds to a region.}

\item{product}{variable in \code{data} which corresponds to a product.}

\item{value}{variable in \code{data} which corresponds to a value.}
}
\value{
dataframe
}
\description{
Calculate economic complexity indicators for multiple years
}
\examples{
\dontrun{

calculate_complexity_time_series(state_data,
years = unique(state_data$year),
region = "location_code",
product = "hs_product_code",
value = "export_value")

}
}
