% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_graphs.R
\name{graph_complexity_product_space}
\alias{graph_complexity_product_space}
\title{Draw Product Spaces}
\usage{
graph_complexity_product_space(
  data,
  country,
  year,
  classification,
  services = FALSE,
  proj = NULL
)
}
\arguments{
\item{data}{export data}

\item{country}{country}

\item{year}{year}

\item{classification}{hs92 or hs12}

\item{services}{should services be included? Default is FALSE which excludes service exports}

\item{proj}{have you already calcualted the projection?}
}
\value{
ggplot
}
\description{
Draw Product Spaces
}
\examples{
\dontrun{
library(dplyr)
data <- read_complexitydata("atlas_economic_complexity12")
graph_complexity_product_space(data, country = "AUS", year = 2024, classification = "hs12")
}
}
