% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_data}
\alias{product_data}
\title{Harmonised System (1992) product data.}
\format{
Tibble with
\describe{
\item{product_id}{The product ID.}
\item{parent_id}{The product ID of the product's parent.}
\item{hs_product_name_short_en}{English short name of the product.}
\item{hs_product_code}{Character representation of the product code at the four-digit HS92 level.}
\item{group_name}{Name of the product group.}
\item{section_name}{Name of the product section. }
}
}
\source{
\url{https://dataverse.harvard.edu/file.xhtml?persistentId=doi:10.7910/DVN/3BAL1O/WQZPYX&version=4.0}
}
\usage{
product_data
}
\description{
A dataset containing the names, product code, group name and section name of products
categorised under the 1992 version of the Harmonised System.
Highest level of disaggregation in this data set is the four-digit HS92
}
\keyword{datasets}
