% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_complexitydata.R
\name{read_complexitydata}
\alias{read_complexitydata}
\title{Read economic complexity data from the complexitydata GitHub repository}
\usage{
read_complexitydata(
  name = NULL,
  refresh = FALSE,
  export_dir = tempdir(),
  .validate_name = TRUE
)
}
\arguments{
\item{.validate_name}{}
}
\description{
Read economic complexity data from the complexitydata GitHub repository
}
